Change GetByIdAsync parameter from int to long
Updated GetByIdAsync method in GenericRepository and IGenericRepository
to accept a long parameter instead of int. This allows handling larger
ID values.

+++ backend/Infrastracture/Base/GenericRepository.cs
@@ -31,7 +31,7 @@ namespace Infrastracture.Base
         #endregion
 
         #region Actions
-        public virtual async Task<T> GetByIdAsync(int id)
+        public virtual async Task<T> GetByIdAsync(long id)
         {
 
             return await _dbContext.Set<T>().FindAsync(id);
